Two changes.
1. Switch to using std::string as the data member in Node.  The traverse
function shouldn't care about formatting of the directory name.  This
prepares the way to adding a prefix to each directory name and
zero-filling the numbers.

2. Use the given base directory (or "." if none was specified) as the
root of the directory tree.  The existing algorithm for generating the
tree always created the root with only one child.  This change causes
the root (base directory) to have one or more children based on the
results of the pseudo-random numbers generated.